separated scene class and started clickable
