+++ Assets/scripts/bedroom.cs.meta
